backend_LogicMed/requirements.txt
Flask
Flask-Cors
python-dotenv
pymongo
azure-ai-inference
azure-core
pytesseract
Pillow
pymupdf
gunicorn